Add teams in same row
