+++ client/src/features/MatchHistory/components/MHMatchDisplay.tsx
@@ -1,7 +1,7 @@
 import { useParams } from "react-router-dom";
 import { MatchDto } from "../../../types/MatchV5";
 import MHBans from "./MHBans";
-import TeamContainer from "./TeamContainer";
+import MatchContainer from "./MatchContainer";
 
 const MHMatchDisplay = ({ matchData }: { matchData: MatchDto }) => {
   const params = useParams();
@@ -13,17 +13,14 @@ const MHMatchDisplay = ({ matchData }: { matchData: MatchDto }) => {
     return `${String(minutes).padStart(2, "0")}:${String(remainingSeconds).padStart(2, "0")}`;
   };
 
-  const blueTeamPlayers = matchData.info.participants.filter(
-    (participant) => participant.teamId === 100
-  );
+  console.log(matchData);
+  const blueTeamPlayers = matchData.info.participants.filter((participant) => participant.teamId === 100);
 
-  const redTeamPlayers = matchData.info.participants.filter(
-    (participant) => participant.teamId === 200
-  );
+  const redTeamPlayers = matchData.info.participants.filter((participant) => participant.teamId === 200);
 
   return (
     <>
-      <div className="flex flex-col gap-2 mt-20 md:p-4">
+      <div className="flex flex-col gap-2 mt-20 md:p-4 w-full">
         <div className="topContainer flex flex-col md:flex-row gap-2 py-4 px-6 bg-gray rounded-md items-center md:items-start md:justify-between">
           <div className="title flex flex-col">
             <h1 className="text-2xl font-bold">Match Details</h1>
@@ -42,12 +39,12 @@ const MHMatchDisplay = ({ matchData }: { matchData: MatchDto }) => {
             </div>
           </div>
         </div>
-        <div className="blueTeamContainer">
-          <TeamContainer team={matchData.info.teams[0]} players={blueTeamPlayers} />
-        </div>
-        <div className="redTeamContainer">
-          <TeamContainer team={matchData.info.teams[1]} players={redTeamPlayers} />
-        </div>
+        <MatchContainer
+          blueTeam={matchData.info.teams[0]}
+          redTeam={matchData.info.teams[1]}
+          blueTeamPlayers={blueTeamPlayers}
+          redTeamPlayers={redTeamPlayers}
+        />
       </div>
     </>
   );
